+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+
+- Added keyboard activation animation to Interaction widget
+
 ## 1.0.1
 
 - Added color argument to getForegroundColor